adaptations for locust 2.29.1
this caused problems for HttpClientTask, where in one E2E test one of the get requests would hang until pytest.timeout kicked in.

replaced `requests` with `geventhttpclient` and it worked. which also is a step to remove `requests` from grizzly.

that in turn made sense to create some util functions that would be used by both HttpClientTask and RestApiUser.

which in turn, again, made sense to move all protocol related (`mq_`, `zmq_`, `async_message_`) helper functions to a new module, togheter with the new `http_` helper functions.
