+++ grizzly/users/servicebus.py
@@ -87,7 +87,8 @@ import zmq.green as zmq
 from grizzly.tasks import RequestTask
 from grizzly.types import GrizzlyResponse, RequestDirection, RequestMethod, RequestType
 from grizzly.types.locust import Environment, StopUser
-from grizzly.utils import has_parameter, has_template, zmq_disconnect
+from grizzly.utils import has_parameter, has_template
+from grizzly.utils.protocols import zmq_disconnect
 from grizzly_extras.arguments import get_unsupported_arguments, parse_arguments
 from grizzly_extras.async_message import AsyncMessageContext, AsyncMessageRequest, AsyncMessageResponse
 from grizzly_extras.async_message.utils import async_message_request
