+++ pyproject.toml
@@ -13,7 +13,7 @@ readme = "README.md"
 license = {text = 'MIT'}
 requires-python = ">=3.8"
 dependencies = [
-    "locust @ git+https://git@github.com/mgor/locust.git@v2.24.1_grizzly",
+    "locust ==2.29.2.dev26",
     "azure-core ==1.30.1",
     "azure-servicebus ==7.12.1",
     "azure-storage-blob ==12.19.1",
@@ -32,7 +32,8 @@ dependencies = [
     "pyotp ==2.9.0",
     "tzdata >=2022.1",
     "websocket-client ==1.7.0",
-    "jinja2-simple-tags ==0.6.1"
+    "jinja2-simple-tags ==0.6.1",
+    "roundrobin ==0.0.4"
 ]
 classifiers = [
     "Development Status :: 4 - Beta",
@@ -104,7 +105,7 @@ docs = [
     "databind ==4.5.1",
     "pytablewriter ==1.2.0",
     "pip-licenses ==4.3.4",
-    "requests ==2.31.0",
+    "requests ==2.32.3",
     "mkdocs ==1.5.3",
     "mkdocs-material ==9.5.15",
     "packaging ==24.0",
